refactor(category): page.tsx を薄く（grouping/presentational/view を分離）
1065行の god component を 230行のオーケストレーターへ。挙動・出力は不変。
- lib/category-groups.ts: grouping/sorting（GROUP_DESCRIPTIONS/sortDocs/groupDocs）
- components/category/CategorySections.tsx: DocCard/DocSection/5 種の試験テーブル
- components/category/CategoryViews.tsx: 資格別 IIFE 分岐を 5 view コンポーネント化
- page.tsx: metadata + データ取得 + サイドバー/転職枠 + view 振り分けのみ
