+++ src/components/category/CategoryViews.tsx
@@ -1,0 +1,209 @@
+import { type ReactNode } from 'react';
+import Link from 'next/link';
+import { getGroupLabel } from '@/lib/doc-classifier';
+import { type DocGroup } from '@/lib/category-groups';
+import { DocCard, DocSection } from '@/components/category/CategorySections';
+
+/** civil-construction-1: primary をテーブル、secondary の年度別を統合 */
+export function CivilConstruction1View({ groups, mobileCareerAd }: { groups: DocGroup[]; mobileCareerAd: ReactNode }) {
+  const guideGroup = groups.find(g => g.title === getGroupLabel('civil-construction-1', 'guide'));
+  const textbookGroup = groups.find(g => g.title === getGroupLabel('civil-construction-1', 'textbook'));
+  const primaryGroup = groups.find(g => g.title === getGroupLabel('civil-construction-1', 'primary'));
+  const secondaryGroup = groups.find(g => g.title === getGroupLabel('civil-construction-1', 'secondary'));
+
+  // secondary を年度別過去問と分野別に分離
+  const secondaryYearDocs = secondaryGroup?.docs.filter(d =>
+    /secondary-(r|h)\d+$/.test(d.slug || '')
+  ) || [];
+  const secondaryTopicDocs = secondaryGroup?.docs.filter(d =>
+    !/secondary-(r|h)\d+$/.test(d.slug || '')
+  ) || [];
+
+  // テキストブックをエリア別にグループ化
+  const TEXTBOOK_AREAS = [
+    { label: '建設機械', min: 100, max: 149 },
+    { label: '測量', min: 150, max: 169 },
+    { label: '解体工事', min: 170, max: 179 },
+    { label: '施工管理・施工計画', min: 200, max: 230 },
+    { label: '工程管理', min: 250, max: 269 },
+    { label: '品質管理', min: 300, max: 320 },
+    { label: '関係法規', min: 400, max: 449 },
+  ];
+  const textbookAreas = textbookGroup ? TEXTBOOK_AREAS.map(area => ({
+    ...area,
+    docs: textbookGroup.docs.filter(d => {
+      const order = d.textbook_order ?? 999;
+      return order >= area.min && order <= area.max;
+    }),
+  })).filter(a => a.docs.length > 0) : [];
+
+  return (
+    <>
+      {guideGroup && <DocSection group={guideGroup} />}
+      {mobileCareerAd}
+      {textbookGroup && (
+        <section>
+          <div className="mb-6">
+            <h2 className="text-2xl font-bold text-gray-900 dark:text-white">{textbookGroup.title}</h2>
+            <p className="text-base text-gray-500 dark:text-gray-400 mt-1">{textbookGroup.description}</p>
+            <span className="text-sm text-gray-400 dark:text-gray-500">{textbookGroup.docs.length} 件</span>
+          </div>
+          <div className="space-y-8">
+            {textbookAreas.map(area => (
+              <div key={area.label}>
+                <h3 className="text-lg font-semibold text-gray-700 dark:text-gray-300 mb-3 border-b border-gray-200 dark:border-gray-700 pb-2">{area.label}</h3>
+                <div className="grid gap-4 md:grid-cols-2 lg:grid-cols-3">
+                  {area.docs.map(doc => (
+                    <DocCard key={doc.slug} doc={doc} />
+                  ))}
+                </div>
+              </div>
+            ))}
+          </div>
+        </section>
+      )}
+      {primaryGroup && (
+        <DocSection
+          group={{
+            ...primaryGroup,
+            title: '過去問',
+            description: '年度別の第1次検定（問題A・問題B）と第2次検定',
+            docs: primaryGroup.docs,
+          }}
+          layout="exam-table"
+          secondaryDocs={secondaryYearDocs}
+        />
+      )}
+      {secondaryTopicDocs.length > 0 && (
+        <DocSection
+          group={{
+            title: '第2次検定 分野別対策',
+            description: '経験記述・施工管理（コンクリート工・土工・品質管理・施工計画）の基礎と過去問',
+            docs: secondaryTopicDocs,
+          }}
+        />
+      )}
+    </>
+  );
+}
+
+/** civil-construction-2: 2級向け、前期/後期テーブル */
+export function CivilConstruction2View({ groups, mobileCareerAd }: { groups: DocGroup[]; mobileCareerAd: ReactNode }) {
+  const guideGroup = groups.find(g => g.title === getGroupLabel('civil-construction-2', 'guide'));
+  const textbookGroup = groups.find(g => g.title === getGroupLabel('civil-construction-2', 'textbook'));
+  const primaryGroup = groups.find(g => g.title === getGroupLabel('civil-construction-2', 'primary'));
+  const secondaryGroup = groups.find(g => g.title === getGroupLabel('civil-construction-2', 'secondary'));
+
+  const secondaryYearDocs = secondaryGroup?.docs.filter(d =>
+    /secondary-(r|h)\d+$/.test(d.slug || '')
+  ) || [];
+  const secondaryTopicDocs = secondaryGroup?.docs.filter(d =>
+    !/secondary-(r|h)\d+$/.test(d.slug || '')
+  ) || [];
+
+  return (
+    <>
+      {guideGroup && <DocSection group={guideGroup} />}
+      {mobileCareerAd}
+      {textbookGroup && <DocSection group={textbookGroup} />}
+      {primaryGroup && (
+        <DocSection
+          group={{
+            ...primaryGroup,
+            title: '過去問',
+            description: '年度別の第1次検定（前期・後期）と第2次検定',
+            docs: primaryGroup.docs,
+          }}
+          layout="exam-table-2"
+          secondaryDocs={secondaryYearDocs}
+        />
+      )}
+      {secondaryTopicDocs.length > 0 && (
+        <DocSection
+          group={{
+            title: '第2次検定 分野別対策',
+            description: '経験記述・施工管理（コンクリート工・土工・品質管理・施工計画）の基礎と過去問（主任技術者視点）',
+            docs: secondaryTopicDocs,
+          }}
+        />
+      )}
+    </>
+  );
+}
+
+/** pe-first-stage: 適性・基礎・専門マトリクス */
+export function PeFirstStageView({ groups }: { groups: DocGroup[] }) {
+  const primaryGroup = groups.find(g => g.title === getGroupLabel('pe-first-stage', 'primary'));
+  return (
+    <>
+      {primaryGroup && (
+        <DocSection group={primaryGroup} layout="pe-first-stage-table" />
+      )}
+    </>
+  );
+}
+
+/** pe-comprehensive-management: ガイド・ピラー・過去問・キーワード索引導線 */
+export function PeComprehensiveView({ groups, mobileCareerAd }: { groups: DocGroup[]; mobileCareerAd: ReactNode }) {
+  const guideGroup = groups.find(g => g.title === getGroupLabel('pe-comprehensive-management', 'guide'));
+  const pillarGroup = groups.find(g => g.title === getGroupLabel('pe-comprehensive-management', 'pillar'));
+  const pastExamGroup = groups.find(g => g.title === getGroupLabel('pe-comprehensive-management', 'pastExam'));
+  const keywordGroup = groups.find(g => g.title === getGroupLabel('pe-comprehensive-management', 'keyword'));
+  const keywordCount = keywordGroup?.docs.length ?? 0;
+
+  // essay-mlit-* 7 記事は 2026-05-18 撤回済み（旧分離ロジック削除）
+  return (
+    <>
+      {guideGroup && guideGroup.docs.length > 0 && <DocSection group={guideGroup} />}
+      {mobileCareerAd}
+      {pillarGroup && <DocSection group={pillarGroup} />}
+      {pastExamGroup && (
+        <DocSection group={pastExamGroup} layout="pe-exam-table" />
+      )}
+      {/* キーワード索引へのナビゲーション（本体は /sitemap-keywords に移動） */}
+      {keywordCount > 0 && (
+        <section>
+          <div className="mb-6">
+            <h2 className="font-serif text-[22px] sm:text-[26px] font-black text-[var(--ink)]">キーワードを探す</h2>
+            <p className="text-[14px] text-[var(--ink-muted)] mt-1">
+              5 管理 × 26 セクションで体系化された全 {keywordCount} キーワードの索引
+            </p>
+          </div>
+          <Link
+            href="/sitemap-keywords"
+            className="group flex items-center gap-4 p-5 rounded-card-content border-2 border-blue-200 dark:border-blue-800 bg-blue-50/50 dark:bg-blue-900/20 hover:border-blue-400 dark:hover:border-blue-500 hover:shadow-card-hover transition-all"
+          >
+            <div className="flex-shrink-0 w-10 h-10 rounded-sm bg-blue-600 dark:bg-blue-500 flex items-center justify-center text-white font-bold text-lg">
+              ≡
+            </div>
+            <div className="flex-1">
+              <div className="font-bold text-lg text-gray-900 dark:text-white group-hover:text-blue-600 dark:group-hover:text-blue-400">
+                キーワードを全件見る（{keywordCount} 件）
+              </div>
+              <div className="text-sm text-gray-500 dark:text-gray-400">
+                文部科学省「総合技術監理 キーワード集 2026」に基づくセクション別索引へ
+              </div>
+            </div>
+            <span className="text-blue-600 dark:text-blue-400 group-hover:translate-x-1 transition-transform" aria-hidden>›</span>
+          </Link>
+        </section>
+      )}
+    </>
+  );
+}
+
+/** pe-construction: ガイド・キーワード・科目×年度マトリクス */
+export function PeConstructionView({ groups }: { groups: DocGroup[] }) {
+  const guideGroup = groups.find(g => g.title === getGroupLabel('pe-construction', 'guide'));
+  const keywordGroup = groups.find(g => g.title === getGroupLabel('pe-construction', 'keyword'));
+  const pastExamGroup = groups.find(g => g.title === getGroupLabel('pe-construction', 'pastExam'));
+  return (
+    <>
+      {guideGroup && <DocSection group={guideGroup} />}
+      {keywordGroup && <DocSection group={keywordGroup} />}
+      {pastExamGroup && (
+        <DocSection group={pastExamGroup} layout="pe-construction-exam-table" />
+      )}
+    </>
+  );
+}
